feat: add more conversion statistics
Storing API response allowed us to perform sums and count on the whole
table and time windows. We propagated this calculation to client as a
tooltip in UI.

This commit introduces breaking changes for dev database (renamed `from`
column to `from_currency` and `to` to `to_currency` to cleanup
conflicting names).